requirements.txt
faiss_gpu==1.7.2
llama_index==0.14.12
modelscope==1.31.0
numpy==2.4.0
pandas==2.3.3
regex==2024.11.6
sentence_transformers==4.1.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
utils==1.0.2
vllm==0.11.0                                                                                                                 "job-fp3urf4fxoxr-work" 14:22 08-Jan-26